attendance/requirements.txt
black==24.4.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
